refactor hdfs backend selector
