+++ fe/fe-core/src/main/java/com/starrocks/qe/BackendSelector.java
@@ -1,0 +1,37 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
+
+package com.starrocks.qe;
+
+import com.starrocks.thrift.TScanRangeParams;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public interface BackendSelector {
+    void computeScanRangeAssignment() throws Exception;
+
+    // TODO(yan): to get which compute nodes have been used during assignment.
+    // public List<ComputeNode> getUsedNodes();
+
+    static <K, V> V findOrInsert(HashMap<K, V> m, final K key, final V defaultVal) {
+        V value = m.get(key);
+        if (value == null) {
+            m.put(key, defaultVal);
+            value = defaultVal;
+        }
+        return value;
+    }
+
+    // weather we can overwrite the first parameter or not?
+    static List<TScanRangeParams> findOrInsert(Map<Integer, List<TScanRangeParams>> m, Integer key,
+                                               ArrayList<TScanRangeParams> defaultVal) {
+        List<TScanRangeParams> value = m.get(key);
+        if (value == null) {
+            m.put(key, defaultVal);
+            value = defaultVal;
+        }
+        return value;
+    }
+}
