+++ fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
@@ -23,13 +23,11 @@ package com.starrocks.qe;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableCollection;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
 import com.starrocks.analysis.DescriptorTable;
 import com.starrocks.catalog.FsBroker;
@@ -118,7 +116,6 @@ import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Deque;
 import java.util.HashMap;
@@ -1995,26 +1992,6 @@ public class Coordinator {
         return newPlan;
     }
 
-    private <K, V> V findOrInsert(HashMap<K, V> m, final K key, final V defaultVal) {
-        V value = m.get(key);
-        if (value == null) {
-            m.put(key, defaultVal);
-            value = defaultVal;
-        }
-        return value;
-    }
-
-    // weather we can overwrite the first parameter or not?
-    private List<TScanRangeParams> findOrInsert(Map<Integer, List<TScanRangeParams>> m, Integer key,
-                                                ArrayList<TScanRangeParams> defaultVal) {
-        List<TScanRangeParams> value = m.get(key);
-        if (value == null) {
-            m.put(key, defaultVal);
-            value = defaultVal;
-        }
-        return value;
-    }
-
     /**
      * This strategy assigns buckets to each driver sequence to avoid local shuffle.
      * If the number of buckets assigned to a fragment instance is less than pipelineDop,
@@ -2022,7 +1999,7 @@ public class Coordinator {
      * Therefore, when there are few buckets (<=pipeline_dop/2), insert local shuffle instead of using this strategy
      * to improve the degree of parallelism.
      *
-     * @param scanRanges The buckets assigned to a fragment instance.
+     * @param scanRanges  The buckets assigned to a fragment instance.
      * @param pipelineDop The expected pipelineDop.
      * @return Whether using the strategy of assigning scanRanges to each driver sequence.
      */
@@ -2116,6 +2093,21 @@ public class Coordinator {
         }
     }
 
+    public ImmutableCollection<ComputeNode> getSelectorComputeNodes() {
+        if (usedComputeNode) {
+            return idToComputeNode.values();
+        } else {
+            return ImmutableList.copyOf(idToBackend.values());
+        }
+    }
+
+    public void updateUsedComputeNodes(List<ComputeNode> nodes) {
+        for (ComputeNode node : nodes) {
+            TNetworkAddress addr = new TNetworkAddress(node.getHost(), node.getBePort());
+            addressToBackendID.put(addr, node.getId());
+        }
+    }
+
     // Populates scan_range_assignment_.
     // <fragment, <server, nodeId>>
     private void computeScanRangeAssignment() throws Exception {
@@ -2132,14 +2124,11 @@ public class Coordinator {
                     fragmentExecParamsMap.get(scanNode.getFragmentId()).scanRangeAssignment;
             if ((scanNode instanceof HdfsScanNode) || (scanNode instanceof IcebergScanNode) ||
                     scanNode instanceof HudiScanNode) {
-                HDFSBackendSelector selector = new HDFSBackendSelector(scanNode, locations, assignment,
-                        ScanRangeAssignType.SCAN_DATA_SIZE);
-                List<Long> scanRangesBytes = Lists.newArrayList();
-                for (TScanRangeLocations scanRangeLocations : locations) {
-                    scanRangesBytes.add(scanRangeLocations.scan_range.hdfs_scan_range.length);
-                }
-                selector.setScanRangesBytes(scanRangesBytes);
+                HDFSBackendSelector selector =
+                        new HDFSBackendSelector(scanNode, locations, assignment, getSelectorComputeNodes(),
+                                forceScheduleLocal);
                 selector.computeScanRangeAssignment();
+                updateUsedComputeNodes(selector.getUsedNodes());
             } else {
                 boolean hasColocate = isColocateFragment(scanNode.getFragment().getPlanRoot());
                 boolean hasBucket =
@@ -2435,16 +2424,6 @@ public class Coordinator {
             }
             fragmentProfiles.get(backendExecState.profileFragmentId).addChild(backendExecState.profile);
         }
-    }
-
-    // For HybridBackendSelector
-    private enum ScanRangeAssignType {
-        SCAN_RANGE_NUM,
-        SCAN_DATA_SIZE
-    }
-
-    private interface BackendSelector {
-        void computeScanRangeAssignment() throws Exception;
     }
 
     // fragment instance exec param, it is used to assemble
@@ -3103,7 +3082,7 @@ public class Coordinator {
                 Long minAssignedBytes = Long.MAX_VALUE;
                 TScanRangeLocation minLocation = null;
                 for (final TScanRangeLocation location : scanRangeLocations.getLocations()) {
-                    Long assignedBytes = findOrInsert(assignedBytesPerHost, location.server, 0L);
+                    Long assignedBytes = BackendSelector.findOrInsert(assignedBytesPerHost, location.server, 0L);
                     if (assignedBytes < minAssignedBytes) {
                         minAssignedBytes = assignedBytes;
                         minLocation = location;
@@ -3121,9 +3100,9 @@ public class Coordinator {
                 }
                 addressToBackendID.put(execHostPort, backendIdRef.getRef());
 
-                Map<Integer, List<TScanRangeParams>> scanRanges = findOrInsert(
+                Map<Integer, List<TScanRangeParams>> scanRanges = BackendSelector.findOrInsert(
                         assignment, execHostPort, new HashMap<Integer, List<TScanRangeParams>>());
-                List<TScanRangeParams> scanRangeParamsList = findOrInsert(
+                List<TScanRangeParams> scanRangeParamsList = BackendSelector.findOrInsert(
                         scanRanges, scanNode.getId().asInt(), new ArrayList<TScanRangeParams>());
                 // add scan range
                 TScanRangeParams scanRangeParams = new TScanRangeParams();
@@ -3150,7 +3129,7 @@ public class Coordinator {
             for (TScanRangeLocations scanRangeLocations : locations) {
                 for (Map.Entry<TNetworkAddress, Map<Integer, List<TScanRangeParams>>> kv : assignment.entrySet()) {
                     Map<Integer, List<TScanRangeParams>> scanRanges = kv.getValue();
-                    List<TScanRangeParams> scanRangeParamsList = findOrInsert(
+                    List<TScanRangeParams> scanRangeParamsList = BackendSelector.findOrInsert(
                             scanRanges, scanNode.getId().asInt(), new ArrayList<>());
                     // add scan range
                     TScanRangeParams scanRangeParams = new TScanRangeParams();
@@ -3171,7 +3150,7 @@ public class Coordinator {
             if (bucketSeqToScanRange != null && !bucketSeqToScanRange.isEmpty()) {
                 for (Map.Entry<Integer, Map<Integer, List<TScanRangeParams>>> entry : bucketSeqToScanRange.entrySet()) {
                     for (TScanRangeLocations scanRangeLocations : locations) {
-                        List<TScanRangeParams> scanRangeParamsList = findOrInsert(
+                        List<TScanRangeParams> scanRangeParamsList = BackendSelector.findOrInsert(
                                 entry.getValue(), scanNode.getId().asInt(), new ArrayList<>());
                         // add scan range
                         TScanRangeParams scanRangeParams = new TScanRangeParams();
@@ -3301,167 +3280,6 @@ public class Coordinator {
 
             addressToBackendID.put(execHostPort, backendIdRef.getRef());
             fragmentIdToSeqToAddressMap.get(fragmentId).put(bucketSeq, execHostPort);
-        }
-    }
-
-    /**
-     * Hybrid backend selector for hive table.
-     * Support hybrid and independent deployment with datanode.
-     * <p>
-     * Assign scan ranges to backend:
-     * 1. local backend first,
-     * 2. and smallest assigned scan ranges num or scan bytes.
-     * <p>
-     * If force_schedule_local variable is set, HybridBackendSelector will force to
-     * assign scan ranges to local backend if there has one.
-     */
-    private class HDFSBackendSelector implements BackendSelector {
-        // be -> assigned scans
-        // type:
-        //     SCAN_RANGE_NUM: assigned scan range num
-        //     SCAN_DATA_SIZE: assigned scan data size
-        Map<ComputeNode, Long> assignedScansPerComputeNode = Maps.newHashMap();
-        // be host -> bes
-        Multimap<String, ComputeNode> hostToBes = HashMultimap.create();
-        private final ScanNode scanNode;
-        private final List<TScanRangeLocations> locations;
-        private final FragmentScanRangeAssignment assignment;
-        private final ScanRangeAssignType assignType;
-        // for SCAN_DATA_SIZE assign type
-        private List<Long> scanRangesBytes = Lists.newArrayList();
-        private final List<Long> remoteScanRangesBytes = Lists.newArrayList();
-        // TODO: disk stats
-
-        public HDFSBackendSelector(ScanNode scanNode, List<TScanRangeLocations> locations,
-                                   FragmentScanRangeAssignment assignment, ScanRangeAssignType assignType) {
-            this.scanNode = scanNode;
-            this.locations = locations;
-            this.assignment = assignment;
-            this.assignType = assignType;
-        }
-
-        /**
-         * Must be set when using SCAN_DATA_SIZE assign type
-         *
-         * @param scanRangesBytes: has the same size with locations
-         */
-        public void setScanRangesBytes(List<Long> scanRangesBytes) {
-            this.scanRangesBytes = scanRangesBytes;
-        }
-
-        @Override
-        public void computeScanRangeAssignment() throws Exception {
-            Preconditions.checkArgument(assignType != ScanRangeAssignType.SCAN_DATA_SIZE
-                    || locations.size() == scanRangesBytes.size());
-            ImmutableCollection<ComputeNode> nodes;
-            if (hasComputeNode) {
-                usedComputeNode = true;
-                nodes = idToComputeNode.values();
-            } else {
-                List<ComputeNode> backends = Lists.newArrayList(idToBackend.values());
-                nodes = ImmutableList.copyOf(backends);
-            }
-            for (ComputeNode computeNode : nodes) {
-                if (!computeNode.isAlive() || SimpleScheduler.isInBlacklist(computeNode.getId())) {
-                    continue;
-                }
-
-                assignedScansPerComputeNode.put(computeNode, 0L);
-                hostToBes.put(computeNode.getHost(), computeNode);
-            }
-            if (hostToBes.isEmpty()) {
-                throw new UserException("Backend not found. Check if any backend is down or not");
-            }
-
-            // total scans / alive bes
-            long avgScansPerBe = -1;
-            if (!forceScheduleLocal) {
-                int numBes = assignedScansPerComputeNode.size();
-                if (assignType == ScanRangeAssignType.SCAN_DATA_SIZE) {
-                    long totalBytes = 0L;
-                    for (long scanRangeBytes : scanRangesBytes) {
-                        totalBytes += scanRangeBytes;
-                    }
-                    avgScansPerBe = totalBytes / numBes + (totalBytes % numBes == 0 ? 0 : 1);
-                } else {
-                    // SCAN_RANGE_NUM
-                    int numLocations = locations.size();
-                    avgScansPerBe = numLocations / numBes + (numLocations % numBes == 0 ? 0 : 1);
-                }
-            }
-
-            List<TScanRangeLocations> remoteScanRangeLocations = Lists.newArrayList();
-            for (int i = 0; i < locations.size(); ++i) {
-                TScanRangeLocations scanRangeLocations = locations.get(i);
-                long minAssignedScanRanges = Long.MAX_VALUE;
-                ComputeNode minBe = null;
-                for (final TScanRangeLocation location : scanRangeLocations.getLocations()) {
-                    Collection<ComputeNode> bes = hostToBes.get(location.getServer().getHostname());
-                    if (bes == null || bes.isEmpty()) {
-                        continue;
-                    }
-                    for (ComputeNode backend : bes) {
-                        long assignedScanRanges = assignedScansPerComputeNode.get(backend);
-                        if (!forceScheduleLocal && assignedScanRanges >= avgScansPerBe) {
-                            continue;
-                        }
-                        if (assignedScanRanges < minAssignedScanRanges) {
-                            minAssignedScanRanges = assignedScanRanges;
-                            minBe = backend;
-                        }
-                    }
-                }
-                if (minBe == null) {
-                    remoteScanRangeLocations.add(scanRangeLocations);
-                    if (assignType == ScanRangeAssignType.SCAN_DATA_SIZE) {
-                        remoteScanRangesBytes.add(scanRangesBytes.get(i));
-                    }
-                    continue;
-                }
-                long scansToAdd = (assignType == ScanRangeAssignType.SCAN_DATA_SIZE ? scanRangesBytes.get(i) : 1);
-                recordScanRangeAssignment(minBe, scanRangeLocations, scansToAdd);
-            }
-
-            if (remoteScanRangeLocations.isEmpty()) {
-                return;
-            }
-
-            Preconditions.checkArgument(assignType != ScanRangeAssignType.SCAN_DATA_SIZE
-                    || remoteScanRangeLocations.size() == remoteScanRangesBytes.size());
-            for (int i = 0; i < remoteScanRangeLocations.size(); ++i) {
-                TScanRangeLocations scanRangeLocations = remoteScanRangeLocations.get(i);
-                long minAssignedScanRanges = Long.MAX_VALUE;
-                ComputeNode minBe = null;
-                for (Map.Entry<ComputeNode, Long> entry : assignedScansPerComputeNode.entrySet()) {
-                    ComputeNode backend = entry.getKey();
-                    long assignedScanRanges = entry.getValue();
-                    if (assignedScanRanges < minAssignedScanRanges) {
-                        minAssignedScanRanges = assignedScanRanges;
-                        minBe = backend;
-                    }
-                }
-                long scansToAdd = (assignType == ScanRangeAssignType.SCAN_DATA_SIZE ? remoteScanRangesBytes.get(i) : 1);
-                recordScanRangeAssignment(minBe, scanRangeLocations, scansToAdd);
-            }
-        }
-
-        private void recordScanRangeAssignment(ComputeNode minBe, TScanRangeLocations scanRangeLocations,
-                                               long addedScans) {
-            TNetworkAddress minBeAddress = new TNetworkAddress(minBe.getHost(), minBe.getBePort());
-            addressToBackendID.put(minBeAddress, minBe.getId());
-
-            // update statistic
-            assignedScansPerComputeNode.put(minBe, assignedScansPerComputeNode.get(minBe) + addedScans);
-
-            // add in assignment
-            Map<Integer, List<TScanRangeParams>> scanRanges = findOrInsert(
-                    assignment, minBeAddress, new HashMap<>());
-            List<TScanRangeParams> scanRangeParamsList = findOrInsert(
-                    scanRanges, scanNode.getId().asInt(), new ArrayList<TScanRangeParams>());
-            // add scan range params
-            TScanRangeParams scanRangeParams = new TScanRangeParams();
-            scanRangeParams.scan_range = scanRangeLocations.scan_range;
-            scanRangeParamsList.add(scanRangeParams);
         }
     }
 }
